LayoutTests/workers/sab/postMessage-clones-growable-expected.txt

Checks that window.postMessage throws DataCloneError for growable SharedArrayBuffers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.postMessage(memory, '*') threw exception DataCloneError: The object can not be cloned..
PASS successfullyParsed is true

TEST COMPLETE
